+++ api/login.js
@@ -45,6 +45,14 @@ exports.get = function(req, res){
             },
             {
                 rel: "Section",
+                domain: "http://127.0.0.1",
+                port: ":8081",
+                path: "register",
+                method : "GET",
+                text : "register"
+            },
+            {
+                rel: "Section",
                 domain: "https://127.0.0.1",
                 port: ":8443",
                 path: "account",