+++ data/command.js
@@ -1,4 +1,5 @@
 var Article = require('./model').model('Article');
+var User = require('./model').model('User');
 
 exports.save_article = function (article, callback) {
     var new_article = new Article();
@@ -10,8 +11,27 @@ exports.save_article = function (article, callback) {
     new_article.save(function (err) {
         if (err) {
             console.log(' - fatal error saving article');
+            callback(err);
+            return;
         }
-        callback();
+        callback(null);
+        return;
+    });
+}
+
+exports.save_user = function (user, callback) {
+    var new_user = new User();
+    new_user.sessionId = user.sessionId;
+    new_user.username = user.username;
+    new_user.password = user.password;
+
+    new_user.save(function (err) {
+        if (err) {
+            console.log(' - fatal error saving user');
+            callback(err);
+            return;
+        }
+        callback(null);
         return;
     });
 }