+++ routes/index.js
@@ -5,3 +5,4 @@ exports.github = require('../routes/github');
 exports.login = require('../routes/login');
 exports.account = require('../routes/account');
 exports.errors = require('../routes/errors');
+exports.register = require('../routes/register');